+++ src/app/myCounseling/page.tsx
@@ -6,7 +6,7 @@ import ProcessList from "@/components/common/ProcessList";
 import pocessData from "../../mocks/kjun/managementRevervationStatue.json";
 import UserReservationItem from "@/components/common/Card/CardItem/UserReservationItem";
 import managementRecent from "../../mocks/kjun/managementRecent.json";
-import UserConsultationStatus from "./UserConultationStatus";
+import UserConsultationStatus from "./UserConsultationStatus";
 const PROCESS_NAME: Record<string, string> = {
   APPLY: "신규예약",
   CONFIRM: "예약확정",
@@ -57,12 +57,12 @@ function MyCounselingPage() {
       <UserConsultationStatus {...data} />
       <ProcessList setIsProcess={setIsProcess} role={"user"} />
 
-      <div className="my-8 h-2 w-full bg-background-secondary"></div>
+      <div className="w-full h-2 my-8 bg-background-secondary"></div>
 
-      <div className="w-full justify-start">
+      <div className="justify-start w-full">
         <div>
-          <h3 className="text-lg pl-1 font-bold">{`${PROCESS_NAME[isProcess]} ${selectData?.length}건`}</h3>
-          <p className="my-1 mb-6 pl-1 text-sm">프라이빗 뱅커가 곧 유선으로 연락을 드립니다.</p>
+          <h3 className="pl-1 text-lg font-bold">{`${PROCESS_NAME[isProcess]} ${selectData?.length}건`}</h3>
+          <p className="pl-1 my-1 mb-6 text-sm">프라이빗 뱅커가 곧 유선으로 연락을 드립니다.</p>
           <ul className="flex flex-col gap-4">
             {selectData &&
               selectData.map(({ reservationId, name, createdAt, type }) => (
